refactor: extract notification subscription hook
